move context back to inner class and remove all TODO
