+++ flink-table/flink-table-planner/src/main/java/org/apache/flink/table/planner/plan/logical/CumulativeWindowSpec.java
@@ -120,9 +120,4 @@ public class CumulativeWindowSpec implements WindowSpec {
                     formatWithHighestUnit(offset));
         }
     }
-
-    @Override
-    public boolean isAlignedWindow() {
-        return true;
-    }
 }
