+++ flink-table/flink-table-planner/src/main/java/org/apache/flink/table/planner/plan/logical/HoppingWindowSpec.java
@@ -120,9 +120,4 @@ public class HoppingWindowSpec implements WindowSpec {
                     formatWithHighestUnit(offset));
         }
     }
-
-    @Override
-    public boolean isAlignedWindow() {
-        return true;
-    }
 }
