+++ flink-table/flink-table-planner/src/main/java/org/apache/flink/table/planner/plan/logical/TumblingWindowSpec.java
@@ -99,9 +99,4 @@ public class TumblingWindowSpec implements WindowSpec {
                     formatWithHighestUnit(size), formatWithHighestUnit(offset));
         }
     }
-
-    @Override
-    public boolean isAlignedWindow() {
-        return true;
-    }
 }
