+++ flink-table/flink-table-runtime/src/main/java/org/apache/flink/table/runtime/operators/window/groupwindow/assigners/CountTumblingWindowAssigner.java
@@ -25,7 +25,7 @@ import org.apache.flink.api.common.typeinfo.Types;
 import org.apache.flink.api.common.typeutils.TypeSerializer;
 import org.apache.flink.table.data.RowData;
 import org.apache.flink.table.runtime.operators.window.CountWindow;
-import org.apache.flink.table.runtime.operators.window.groupwindow.context.WindowContext;
+import org.apache.flink.table.runtime.operators.window.groupwindow.internal.InternalWindowProcessFunction;
 
 import java.io.IOException;
 import java.util.Collection;
@@ -47,7 +47,7 @@ public class CountTumblingWindowAssigner extends GroupWindowAssigner<CountWindow
     }
 
     @Override
-    public void open(WindowContext<?, CountWindow> ctx) throws Exception {
+    public void open(InternalWindowProcessFunction.Context<?, CountWindow> ctx) throws Exception {
         String descriptorName = "tumble-count-assigner";
         ValueStateDescriptor<Long> countDescriptor =
                 new ValueStateDescriptor<>(descriptorName, Types.LONG);
