--- flink-table/flink-table-runtime/src/main/java/org/apache/flink/table/runtime/operators/window/groupwindow/context/OnMergeContext.java
@@ -1,35 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.flink.table.runtime.operators.window.groupwindow.context;
-
-import org.apache.flink.api.common.state.MergingState;
-import org.apache.flink.api.common.state.StateDescriptor;
-import org.apache.flink.table.runtime.operators.window.Window;
-import org.apache.flink.table.runtime.operators.window.groupwindow.triggers.Trigger;
-
-/**
- * Extension of {@link TriggerContext} that is given to {@link Trigger#onMerge(Window,
- * OnMergeContext)}.
- */
-public interface OnMergeContext extends TriggerContext {
-
-    /** Merges the state of the given state descriptor. */
-    <S extends MergingState<?, ?>> void mergePartitionedState(
-            StateDescriptor<S, ?> stateDescriptor);
-}
