+++ flink-table/flink-table-runtime/src/main/java/org/apache/flink/table/runtime/operators/window/groupwindow/triggers/ElementTriggers.java
@@ -23,8 +23,6 @@ import org.apache.flink.api.common.state.ReducingState;
 import org.apache.flink.api.common.state.ReducingStateDescriptor;
 import org.apache.flink.api.common.typeutils.base.LongSerializer;
 import org.apache.flink.table.runtime.operators.window.Window;
-import org.apache.flink.table.runtime.operators.window.groupwindow.context.OnMergeContext;
-import org.apache.flink.table.runtime.operators.window.groupwindow.context.TriggerContext;
 
 /**
  * A {@link Trigger} that fires at some point after a specified number of input elements have
